Rename DataContextResourceIdentifier to DataContextKey, with inheritance from AllowedKeysDotDict instead of OrderedKeysDotDict Move OrderedKeysDotDict to base_resource_identifiers.py Move tests for OrderedKeysDotDict to test_base_types.py Resolve naming and imports throughout the codebase.
