+++ great_expectations/data_context/types/__init__.py
@@ -11,10 +11,12 @@ from .metrics import (
 from .configurations import (
     DataContextConfig
 )
+from .base_resource_identifiers import (
+    DataContextKey,
+    OrderedDataContextKey,
+)
 from .resource_identifiers import (
-    DataContextResourceIdentifier,
     DataAssetIdentifier,
-    BatchIdentifier,
     ExpectationSuiteIdentifier,
     ValidationResultIdentifier,
 )
