+++ great_expectations/data_context/util.py
@@ -130,8 +130,9 @@ def safe_mmkdir(directory, exist_ok=True):
         if e.errno != errno.EEXIST:
             raise
 
-# TODO : Consider moving this into types.resource_identifiers.DataContextResourceIdentifier.
+# TODO : Consider moving this into types.resource_identifiers.DataContextKey.
 # NOTE : We **don't** want to encourage stringification of keys, other than in tests, etc.
+# TODO : Rename to parse_string_to_data_context_key
 def parse_string_to_data_context_resource_identifier(string, separator="."):
     string_elements = string.split(separator)
