+++ great_expectations/types/__init__.py
@@ -4,6 +4,7 @@ from .base import (
     DotDict,
     RequiredKeysDotDict,
     AllowedKeysDotDict,
+    OrderedKeysDotDict,
 )
 
 from .expectations import (
