+++ tests/types/test_resource_identifiers.py
@@ -5,8 +5,11 @@ logger = logging.getLogger(__name__)
 from six import PY2, string_types
 import sys
 
+from great_expectations.data_context.types.base_resource_identifiers import (
+   OrderedKeysDotDict,
+)
+
 from great_expectations.data_context.types.resource_identifiers import (
-    OrderedKeysDotDict,
     DataAssetIdentifier,
     ExpectationSuiteIdentifier,
     ValidationResultIdentifier,
@@ -14,67 +17,6 @@ from great_expectations.data_context.types.resource_identifiers import (
 from great_expectations.data_context.util import (
     parse_string_to_data_context_resource_identifier
 )
-
-def test_OrderedKeysDotDict_subclass():
-    # NOTE: Abe 2019/08/23 : The basics work reasonably well, but this class probably still needs to be hardened quite a bit
-    # TODO: Move this to types.test_base_types.py
-
-    class MyOKDD(OrderedKeysDotDict):
-        _key_order = ["A", "B", "C"]
-        _key_types = {
-            "A" : string_types,
-            "B" : int,
-        }
-
-        # NOTE: This pattern is kinda awkward.
-        # It would be nice to ONLY specify _key_order
-        # Instead, we need to add these two lines at the end of every OrderedKeysDotDict class definition
-        # ... There's probably a way to do this with decorators...
-        _allowed_keys = set(_key_order)
-        _required_keys = set(_key_order)
-
-    MyOKDD(**{
-        "A" : "A",
-        "B" : 10,
-        "C" : "C",
-    })
-
-    #OrderedKeysDotDicts can parse from tuples
-    MyOKDD("a", 10, "c")
-
-    #OrderedKeysDotDicts coerce to _key_types by default
-    assert MyOKDD("10", "10", "20") == {
-        "A" : "10",
-        "B" : 10, # <- Not a string anymore!
-        "C" : "20",
-    }
-
-    with pytest.raises(ValueError):
-        assert MyOKDD("a", "10.5", 20)
-
-    #OrderedKeysDotDicts raise an IndexError if args don't line up with keys
-    with pytest.raises(IndexError):
-        MyOKDD("a")
-
-    with pytest.raises(IndexError):
-        MyOKDD("a", 10, "c", "d")
-
-
-def test_OrderedKeysDotDict__recursively_get_key_length():
-
-    class MyOKDD(OrderedKeysDotDict):
-        _key_order = ["A", "B", "C"]
-        _key_types = {
-            "A" : string_types,
-            "B" : int,
-        }
-        _allowed_keys = set(_key_order)
-        _required_keys = set(_key_order)
-
-    assert MyOKDD._recursively_get_key_length() == 3
-    assert DataAssetIdentifier._recursively_get_key_length() == 3
-    assert ValidationResultIdentifier._recursively_get_key_length() == 5
-
 
 def test_DataAssetIdentifier():
 
@@ -88,7 +30,6 @@ def test_DataAssetIdentifier():
     assert my_id.to_string() == "DataAssetIdentifier.A.B.C"
 
 # NOTE: The following tests are good tests of OrderedDotDict's ability to handle abbreviated input to __init__ when nested
-
 def test_ValidationResultIdentifier__init__totally_nested():
     my_id = ValidationResultIdentifier(
         coerce_types=True,
